scripts/ai-daily/requirements.txt
feedparser==6.0.11
requests==2.32.3
anthropic==0.28.0
pyyaml==6.0.2
markdown==3.6
